refactor: move ProtectedRoute to separate file
